chore(dropdown): update files in src instead of v1 folder
