+++ src/v1/components/Menu/Menu.test.tsx
@@ -2,14 +2,14 @@ import { mount, shallow } from 'enzyme';
 import React from 'react';
 
 import { Menu } from './Menu';
-import { MenuItemInfoSchema } from './MenuContent/MenuContent';
-import { MenuSearchResultContent } from './MenuSearchResultContent/MenuSearchResultContent';
 import {
 	menuItems,
 	menuItemsWithDivider,
 	menuItemsWithIcons,
 	menuItemsWithSearch,
-} from './__mocks__';
+} from './Menu.mocks';
+import { MenuItemInfoSchema } from './MenuContent/MenuContent';
+import { MenuSearchResultContent } from './MenuSearchResultContent/MenuSearchResultContent';
 
 describe('<MenuItem />', () => {
 	it('Should be able to render', () => {
